GeNITA/requirements.txt
torch
torchvision
Pillow
tqdm
matplotlib
kagglehub
pyyaml
bitsandbytes
accelerate>=0.26.0
numpy
transformers
typing-extensions